Enhance -f option
The -f option by itself asks the server for all attributes. This
mod makes it so if -f is specified more than once, the server is
asked for only those attributes that would be needed if -f was not
given at all. This can greatly reduce the data passed from the
server and processed by qstat.